significantly reduce size of Docker image using multi-stage builds.
Dependencies are compiled and installed in the first stage into a virtual environment, made with python's standard venv. In the second stage, a slim python docker image is used (e.g., no compilers), and the virtual environment is copied into the second stage. The result is a docker image with a size of 341 MB (uncompressed) vs 1.13 GB before this commit.