Apply search and category filters together
This commit fixes the search and category filters getting
misaligned, e.g. enabling and disabling a category filter leading
to the search filter no longer applying. It also fixes a bug where
deselecting all category filters would not result in all aliases
being visible again (unless filtered out by the word filter, of
course).

To fix this, I split the filter code in two parts: updating a model
containing the current filters (i.e. `activeWordFilter` and
`activeCategoryFilters`), and a function that reads that model and
iterates over all aliases currently present in the DOM and chooses
whether to display them based on that model.

This means that filters can be updated independently, with the
resulting view always being consistent regardless of the order the
filters are applied in.

I also removed the duplicate administration of which aliases and
labels are currently available; instead, the `applyFilters`
function takes the DOM as the source of truth for available
aliases and labels. Ideally it would use the same data there as is
used for displaying those aliases in the first place, but that
would be a larger refactoring for which now is not the time.