Only hide add-on banner with the add-on installed
This required a bit of code archeology, so I'm not sure if I
correctly interpreted the original intentions of the code, but it
seems like we used to only have a banner suggesting the
installation of the add-on, which then got updated to hide
.dashboard-banners when the add-on was installed. Over time, more
banners were added ("download Firefox", "upgrade to Premium"), but
those were all stashed inside .dashboard-banners, which got
hidden when the add-on was installed. Additionally, it seems the
banners are supposed to appear on the page a little bit after page
load? Or maybe that is to give the add-on time to load.

Either way, I now opted to revert the code that hides
.dashboard-banners if the add-on is installed (and thus now only
makes it appear after a delay of half a second). Instead, the
add-on banner specifically is hidden using CSS only, using a
selector that only matches when the add-on has set
`data-addon-installed` to true in <firefox-private-relay-addon>.