Remove "Copied!" tooltip after it's hidden
Where the "Copied!" tooltip would previously only be removed when
another alias was copied, it is now removed as soon as it's hidden
from view. This prevents the animation from being replayed every
time the alias is hidden from view, then shown again (i.e. when a
filter is applied, then unapplied).