Start first new landing page design implementation
This adds a /newlanding/ page that implements one design ("A") for
a new landing page. The page is split up in several partials to
make it easier to avoid merge conflicts when working on the
implementation simultaneously.

It also moves a few globally-applied styles to only apply to pages
wrapped in .c-body-wrapper.t-dark, because they conflicted with the
design.

Unfortunately it seems there's still some global styling
that I've broken, as the links in the header and footer aren't
styled the way they're supposed to anymore.