Position "Copied!" label to the right of the alias
It used to be absolutely positioned to the right of the alias,
presumably to prevent the underline when hovering the button from
also underlining the copy icon by taking it out of the flow.
However, although the label was positioned to the right of the
label, it did so based on the center point of the label, resulting
in it overlapping the alias if the label was longer (like in other
languages).

This reworks that to just only place the underline under the alias
when hovering, and just positioning the "Copied!" label to the
right of the copy icon. The copy icon also no longer disappears.