Landing Page - Priority 1 sections (Hero, How it works, Pricing, Use Case accordion) (#1177)
* Start first new landing page design implementation

This adds a /newlanding/ page that implements one design ("A") for
a new landing page. The page is split up in several partials to
make it easier to avoid merge conflicts when working on the
implementation simultaneously.

It also moves a few globally-applied styles to only apply to pages
wrapped in .c-body-wrapper.t-dark, because they conflicted with the
design.

Unfortunately it seems there's still some global styling
that I've broken, as the links in the header and footer aren't
styled the way they're supposed to anymore.

* Landing page (Hero Section, How it works, Pricing) Priority 1 Sections

* add brands, organize grid

* text align start for how it works sec

* add use case section

* fix accordion

* add strings, move scss into banner.scss

* add localized strings

* add ui fixes

* lint fixes

* update protocol tokens and font sizes

* add font mixin

* apply maxx suggestion, add image and full width to accordion

* add box shadow feedback to accordion

Co-authored-by: Vincent <git@vincenttunru.com>