Only hide add-on-related banners by default
Interpreting code that was there, it seems like the entirety of
div.dashboard-banneres was hidden by default and only shown after
we could be sure whether the add-on was installed, presumably to
show the banners related to either installing the add-on or
installing Firefox.

This was problematic because there were other banners in there,
specifically the banner to register your own custom subdomain. By
hiding that by default, anchor links pointing to that banner did
not work. And since that banner should be shown regardless of
whether the add-on is installed or not, I modified the code to show
that banner by default and only hide the add-on-related banners
until we've determined whether the add-on is installed.