Move the subdomain extension to the next line
This ensures maximum space for the user's chosen subdomain. Note
that we're currently abusing FTL, setting the `subdomain` variable
in `modal-domain-register-success` to an empty string and then just
stuffing the actual chosen subdomain in front of it. This will
break in languages where the chosen subdomain should not be the
first word, but is the result of not loading the language strings
client-side and thus not being able to substitute in variables
generated on the client-side. This was already the case, and this
commit keeps it that way.