Add ability to add translations in-repo
To avoid overwhelming our volunteer translators, we want to submit
as many strings as possible in single batches (otherwise they will
get a notification for every string we add). However, while we
wait for them to be added to the localisation repository, the site
would display `???` until they got merged in.

With this change, we'll be able to simply add them to
`pendingTranslations.ftl` while we wait for them to be added to the
localisation repository.