Only show NPS banner for accounts over 3 days old
This sets a cookie when the user visits the website while logged
in, containing a timestamp. Then if the timestamp in that cookie is
more than three days ago, the NPS banner is shown.

This means the user could see the banner again if they're viewing
the site on another browser. That is accepted for now.