Hide add-on banners using .is-hidden-with-addon
This is a class newly introduced in the extension that will be
injected before JS is executed. This means that we can remove
JavaScript that runs after a 500ms delay and only shows elements
afterwards if it did not find the add-on, and instead can show it
directly on page load (if the add-on is not installed), avoiding
layout shifts.