Allow manually creating custom aliases
Users are reporting that it is unclear to them how to use their
custom subdomain. While it is recommended that they use random
aliases when easy to do so (i.e. when on the website), giving them
a flow to create custom aliases gives us hooks to provide further
education on what custom aliases can be useful for.