Let Django (/Whitenoise) serve the React app
Unfortunately to do this, we have to build the static files of the
React app, rather than using Next.js's built-in dev server. And
with Next.js, builds are always production builds. This means that
it's quite a bit slower than running the dev server, and means we
cannot use different .env files in different environments.

The latter is worked around by using a runtime config (see
https://nextjs.org/docs/api-reference/next.config.js/runtime-configuration
), the former is accepted for now.

One other solution I looked into is making Django act as a reverse
proxy when running locally, proxying into the dev server.
Unfortunately, that led to stylesheets and JS not being loaded,
presumably because their imports were pointing to a different port?
That might be something to investigate in the future.