fix MPP-1512: send ga event for any premium user with "clicked-purchase" cookie
FXA sends the subscription state change event directly to our back-end
as soon as the user completes a purchase, and we process those events as
soon as we get them.

So, our check for "newly_premium" never evaluated to true, because the
event had already updated the profile to premium by the time the user
got back to the profile page.

So instead, always send the "user_purchased_premium" event to GA when a
user with a "clicked-purchase" cookie lands back on the profile page.

Additionally, move the code that updates date_subscribed to the
event-handling code.