fix #1153: use spamVerdict to automatically block spam
If spam is sent to and thru a Relay alias, MTAs sometimes send bounces
back to Relay with a message like: "Message considered as spam ..."

So, when we receive a bounce for a user, if the bounce details includes
the term 'spam', instead of setting a bounce pause on the account,
set a new auto_block_spam field.

Then, when we receive any email for a user with auto_block_spam, and the
SES spamVerdict status is FAIL, don’t forward the email.

This should automatically stop us from forwarding any more spam to a
user after their MTA has bounced a first spam message back.