Also link to Premium with no known country code
When the user's browser's language was set to e.g. just `fr`
instead of `fr-FR` when they created their Firefox account, we used
to err on the side of not showing them the Premium promo.

With this change, we'll still show the link, as long as the
language code is equal to the country code of a country in which
Premium is available. In the rare case where Premium isn't
available after all, the user will be notified of that on /premium.