Always log new subscription after "Purchase" click
This applies the change from
b283d971e8d2a42dd131bafe7fa02aa6a9be609f to the profile_refresh
endpoint. Essentially, if the user has clicked a button to get
Premium, and a refresh of profile data shows that the user has
Premium, we count that as a new subscription - after all, those
buttons should not be shown to users who already have Premium.