Set up infra for a mocked API backend
Using Mock Service Worker, we can build a version of the front-end
with the API mocked out. This mock can be re-used in e.g. tests and
storybooks. And given that it's relatively hard to spin up a new
server running the back-end, this allows us to at least do that for
the front-end, so we can demonstrate e.g. new interactions before
merging them into main.