Sentry:ignore request.summary, DisallowedHost logs
`request.summary` logs at error level for uncaught exceptions, and Sentry
creates an event with the text returned to the caller. However, there is
already an event for the uncaught exception, which includes a traceback
that is useful for debugging. Ignore this logger to avoid the duplicate
and less-useful event.

`django.security.DisallowedHost` is raised when the Host header is not
in settings.ALLOWED_HOST, and returns a 400 Bad Request. In deployments,
these have all been security scan attempts, with a variety of hostnames
or IP addresses. No further action is needed by the team, so ignore the
logger and omit the events from Sentry.