Correctly communicate authn status to add-on
I'd previously assumed that all attributes on
<firefox-private-relay-addon> were set by the add-on and read by
the website (as opposed to <firefox-private-relay-addon-data>,
whose attributes are set by the website and read by the add-on).
However, apparently that is not true for data-user-logged-in: that
attribute is set by the website and read by the add-on. What's
more, it's set to the strings "True" or "False" (like booleans in
Python), rather than "true" or "false".

That means that when the user logged out, the add-on did not know
about it, when running the React front-end. This fixes that.