Make sure the Tips are at the bottom of the page
When the viewport was taller than the content, i.e. when there was
a large amount of whitespace between the alias list and the footer,
the Tips element would be stuck to the bottom of the alias list,
with the whitespace below it, rather than stuck to the top of the
footer, with the whitespace above it.