Switch from cache to workspace
The CircleCI workspace feature is used to save files that are used in
later jobs. This is a better fit for test results than the cache, which
is used to cache files between test runs.

Use /tmp/workspace/test-results, in case we want to use the workspace
for other items, like docker.tar.