Set asserted value in test
Rather than having a test depend on a value implicitly set in the
function, the test now explicitly passes the value it's matching on
into the mock creator. This can avoid the test accidentally passing
if the mocks are changed in the future. (For example, for this test
specifically, it's checking whether the string `some` is in a
mocked email's subdomain, but it would also succeed if that string
was in the address.)

Obviously in this one instance, it's not super important, but I
thought it could be useful to clarify what I meant by my comment in
https://github.com/mozilla/fx-private-relay/pull/1677#discussion_r840517332.