More accurately align "What's new" popup's chevron
We used to somewhat guess where the chevron (attaching the popup to
the button) should go. This change moves the right-hand side of the
menu to be aligned with the right-hand side of the trigger button,
so that the chevron is more accurately attached if it's positioned
a little from the right-hand side.