Add check_health management command
"./manage.py check_health <path_to_healthcheck.json>" reads the
(optional) file written by process_emails_from_sqs, checks if the
timestamp is recent (120 seconds by default), and exits with a code of 1
if too old. This can be used as a Kubernetes liveness check of this and
other processes without a webserver.