Properly mock logout
On the production website, a logout redirects the user to Firefox
Accounts, thereby inducing a full page refresh. However, when
mocking that logout flow locally, we remain in the context of the
Single-Page App, and therefore user data remained cached in memory.
This change forces a revalidation of user data after clearing the
session key, which then results in that data no longer being
available and thus the user's logged-out status being properly
recognised.

It also more thoroughly detects whether the user is logged in on
the homepage, by not only checking if user data is available, but
also that refetching it did not result in an error (which it does
if the user is logged out).