Fix ever-expanding messages cookie
django-allauth adds login and logout messages to the `messages`
cookie, but with Django no longer rendering the view, it is never
reading out and clearing that cookie. Since the cookie is
HTTP-only, the client can't read it out or clear it either. Hence,
here we tell it not to set those messages by overriding its message
templates, as suggested in
https://django-allauth.readthedocs.io/en/latest/advanced.html#messages.