Remove cached data when logging out
Although we now properly detect that a user is logged out after
mocking logout, the user's data would still be cached, such as the
fact that they're a Premium user, causing the header to still be
in the Premium theme.

With this change, user data is also cleared.

Note that this only affected when running using the dev server, as
authentication in production involves full page reloads and hence
the clearing of all in-memory data.