check request.COOKIES for session cookie
We want to keep Session and Authentication middleware AFTER Whitenoise,
so we're not running the session & auth logic on every request for
static files. But we also want to redirect logged-in users to
/accounts/profile/ BEFORE they see a flash of not-logged-in content on
the React index.html.

So, rather than check for request.user, which depends on the Session and
Authentication middleware running first, we can check if there's any
session cookie in the request at all - which indicates the user is
signed-in.

This fixes the flash-of-unauthenticated-content with a light-weight
middleware that is easier on the requests for static files.