Add generic Python jobs, and black test
Add a generic python_job that checks out the project, runs a command
that may fail, and upload results if it generates them. This can be used
for other formatting and linting checks later.

Use the python_job for a black formatting check that is allowed to fail for now.