Remove code made redundant by the React UI
This removes most of the views rendered by Django, linters applying
only to those files, and the infra to compile Sass. It preserves
the package.json file at the root, primarily for our Heroku
configuration (to set the Node and npm versions, and the commands
to run to build the website - i.e. heroku-postbuild), but also to
list the one in `frontend` as an npm workspace, leading to an npm
install in the root also running frontend/'s postinstall scripts
and thereby set up the Git hooks.

I've left the routes to the Django-rendered FAQ and home pages in
place (even though those views are now empty), because they're
still used to generate links. It might be more sensible to just
replace those with direct links, but I'll let people with more
Django experience decide on that.

I've also left the template tags in place, in case they still come
in useful when e.g. rendering the forwarded email.