Add basic 429 template
Add a basic, unstyled, untranslated 429 template, used by
django-allauth.

django-allauth has rate limiting for different actions, including
account signup, and renders `429.html` if the request is rate limited.
In the current version 0.50.0, rate limiting doesn't apply to GET
requests, and we use a GET to start FxA signin, so it doesn't apply to
Relay at this time. However, if it changes in the future, this ensures
users will get a basic HTML page rather than a server error.