Don't send labels to the back-end when disabled
Although we did store modified labels in the add-on, I'd forgotten
to then skip the code path to also send those labels to the
back-end. While I've already updated the back-end to then drop
those labels, it's good to also not even send them there if
disabled by the user, so that the user doesn't have to trust the
back-end to do the right thing.