Use a proper empty exception handler
I'd assumed that an async function that didn't return nor was
`await`ed wouldn't throw, but it looks like there's some machinery
in the dev server that does bubble up the error (which would be
helpful if we didn't intend to swallow the exception!).

Thus, I've now used an explicitly-empty exception handler.