Ensure labels are saved even with add-on present
The mistake I made before was to think that `useLocalLabels` would
return `null` for the local labels if server-side storage of labels
was disabled. However, that was not the case: it would return
`null` if the user did not have the add-on installed. Otherwise, it
would simply return an empty array.

The fix was to only store labels locally (and avoid sending them to
the server) if the user has actually disabled server-side label
storage.