Block release builds on untranslatable strings
Rather than just logging this as a warning in debug mode, with this
change it always gets logged when some strings have not been merged
into the submodule yet, and builds are actually broken off if it's
a build in CircleCI triggered by a new Git tag (i.e. it's a build
that will go to stage and could eventually end up in production).