Don't retry auth requests when logged out
We make requests to /api/v1/profiles/ and /api/v1/users/ to
determine whether the user is currently logged in. When the user is
not logged in, those requests will return a 403, as expected.
However, that means those requests do not need to be retried, which
is SWR's default behaviour for failed requests.

The `onErrorRetry` option is documented at
https://swr.vercel.app/docs/error-handling#error-retry.