Deduce a mask's type from the API response
The /relayaddresses/ and /domainaddresses/ now include a
`mask_type` field, meaning we no longer need our
`markAsRandomAlias` and `markAsCustomAlias` functions that add a
`type` property.

Note that the add-on will be updated to inject `mask_type` as well,
but I left in some compatibility code to ensure that the order in
which we release the website and add-on does not matter.