Make the "Copied!" notification visible again
It was made invisible in 2f07551e0d0903be38c0347e7396be709f748caf,
because the notification was located outside the `.copy-controls`,
and anything outside it was hidden to prevent long mask addresses
from breaking the layout.

Thus, with this change, I've simply moved the notification to
overlap the copy button and be positioned in the middle of it.