Prevent Premium tooltip from expanding the slider
When free users try to select promotional email blocking, we pop
up a tooltip that explains that they will need to upgrade to
Premium for that.

However, because the tooltip element was located inside the
slider's track, useSlider counted it as part of the slider, and
clicking it would change the slider value.

To solve this, I wrapped an element (.track-wrapper) around the
track, that contains both the track and the tooltip. Everything can
then be positioned relative to that wrapper instead of relative to
the track, without that track being considered the slider.

(Most of this complexity is due to the weird hybrid we have of both
a slider and a button that pretends that it is part of the slider,
but actually is not. Perhaps at some point we'll want to split up
<PromotionalTrackStop> into the "real" one that behaves as normal
for Premium users, and is simply hidden for non-Premium users, and
a button that *looks* like the real one, but actually is completely
separate from the slider.)