Split ghost and actual Promotional track stops
Moving the Promotional Track Stop out of the regular track worked
well to make the ghost button not be part of the slider, but it
was somewhat awkward for the actual button (i.e. for Premium
users), where it *should* be part of the slider. This resulted in
people being able to select label text of just the Promotional
track stop, dragging it, and then afterwards not being able to
select Promotional email blocking anymore — presumably due to
react-aria not being able to apply regular track stop behaviour to
it.

This restores the Promo-blocking track stop as a normal one for
Premium users, only keeping what's now called
<PromotionalTrackStopGhost> for the ghost button.