Make full Promotional ghost button clickable
To look like a regular track stop, the ghost button was resized to
the small circle of a track stop, with the label displayed outside
of it. However, that meant that only that small circle was the
actual, clickable button.

Now, I added a <span> that looks like the track stop, and made the
button span a full one-third of the slider, and the same height, so
that it can fully encompass both the track stop and its label.