Remove front-end-specific feature flags
We're replacing them with a more sophisticated mechanism that does
not require recompiling the front-end, and allows gradually
enabling features for specific groups of users.

Note that this does enable the hitherto gated feature
`generateCustomAliasSubdomain` (a popover when clicking your
subdomain explaining how to create a custom mask). Additionally, it
makes the interview recruitment banner gated by the new-style
feature flags, which doesn't have a flag for that feature yet. That
means it will in effect be disabled for now.