Add @layer to ensure website style overrides
Apparently in Chrome, the stylesheet injected by the extension
isn't injected after the website stylesheet, meaning that add-on
styles with the same specificity as website styles do not override
the latter. By moving them into a layer [1], they will. (As long as
the add-on styles are not in a layer, they will override styles
that are, but I've also added a new layer `add-on-overrides` that
the add-on can use to explicitly override the website's
`add-on-overridable` layer.)

[1] https://developer.mozilla.org/en-US/docs/Web/CSS/@layer