Recruit new interview participants
Note that I moved most of the logic to determine whether the
interview recruitment banner should be shown to people into
<TopMessage>; the advantage is that if it is not relevant to a
user, we can show the CSAT survey. (But we won't show the CSAT
survey if the user was shown the recruitment banner first;
that'd be a bit much.)

I also removed the NPS survey, since we never want to show that
anymore.