List tracker blocking if Premium's unavailable
In countries were Relay Premium is not available, tracker blocking
would not be listed under the plans as an available feature. The
reason for this is that if Premium was not available, we weren't
even passing `runtimeData` to the `<Plans>` component (assuming
that it wouldn't be necessary if there were no plans to select).
However, that data is also needed to determine whether the flag is
active already.