Change style class of copy button to focus-visible
The default outline that should appear when navigating
with a keyboard does not show up when the focus lands on
the copy button. As a workaround, PR #2218 adds a border
that changes color on focus. However, :focus doesn't
follow the same behavior as the default styling; the
former applies styling when the focus is from the keyboard
OR the mouse. This commit changes the styling class to
:focus-visible, which only applies the styles to a button
if focused via keyboard.