Allow up to 10 JS Dependabot PRs
If one library with a bunch of packages (e.g. ESLint) releases
updates to several packages at once, our update queue can get
congested. The dependency update tasks on Monday morning isn't
too much work to handle, so I think it's OK for it to be larger
to avoid this problem.