Add News entry for the intro pricing end
Unfortunately I haven't been able to get the Countdown Timer to
update dynamically yet in the news entry, which I'm guessing might
be the result of some interaction between it and the different
React Aria components? For now, I expect that a static counter is
better than not having it at all.