Do not retry requests on server errors
When the server returns an error response, SWR would automatically
retry the request with exponential back-off. Since this might play
a role in the user getting rate-limited (even if just by APIs that
we're calling on the server), this disables that mechanism on
server errors. If e.g. the user's connection drops out, the default
retry mechanism still applies.

Note that requests to `/users/` and `/profiles/` still only retry
on 403 errors, because they already had an exception for just that.
I left that in just in case we revert this change later - we'll
still want to preserve that.