Fix CircleCI job failing when pushing tags
The CircleCI job that tests migrations would try to compare against
the branch that CircleCI was running against. However, when pushing
a tag, there is no known branch (i.e. $CIRCLE_BRANCH is unset,
whereas $CIRCLE_TAG is). To check out the actual relevant commit
regardless of whether we're running for a branch of a tag, checking
$CIRCLE_SHA1 should work.