Redirect user to /premium if they can't buy phones
I had to re-order the components in phone.page.tsx a bit; the
reason for this is that if a user has purchased their phone plan
in another country, but currently is in a country where it's not
available, they should still be able to see their dashboard. Hence,
I moved the dashboard to the top, and moved the
isPhonesAvailableInCountry() check to after that.