Locale-specific price formatting
For the current Premium plan, the price formatting is hardcoded by
the back-end. However, correct formatting is locale-dependent:
some use comma's to separate decimal numbers, some periods, some
put the currency symbol at the start, some at the end, etc.
When we're switching to newly added endpoint data, we might as well
outsource the formatting of pricess to the browser (via its
Intl.NumberFormat API).

The only thing that makes me slightly nervous is that we're dealing
with money using floating-point numbers, but that's what
Intl.NumberFormat expects, and as long as we don't do math with it,
we should be fine.