Cleaner workaround for overlay bug
The actual bug appears to be
https://github.com/adobe/react-spectrum/issues/3490

There's still a failing test where the test suddenly no longer
appears to be able to type into a field. If we actually want to
merge this, we'll have to look into that. (It might be due to my
introducing `mergeProps` where it wasn't present before.)