Prevent lockfile getting out-of-date
While CircleCI supposedly uses `npm ci` to install dependencies,
which should warn if the lockfile does not match what is specified
in package.json, a mismatch did slip in without failing the build.
My hypothesis was that this was due to dependencies being cached
from a previous run, but that did not appear to influence it (a
local `npm ci` *also* did not error). That said, it is still a
good idea to ensure the cache is not re-used if our dependencies
change.