Use correct country code for Sweden
We accidentally used the language code (sv) rather than the country
code (se) for Sweden. This did not cause any problems, because the
mapping from country and language to a plan made the same mistake,
so it evened out. Still, probably a good idea to use the correct
code.

To locally verify that it still works, set your Accept-Language
to sv-SE, visit the locally running website, and verify that the
upgrade links still point to the correct plan.