Change vpn download  link
If a VPN client has come from the client to purchase the bundle then if we send the user to vpn.mozilla.org first, guardian can detect this and can redirect the user back into the client flow. If the user has not come from the VPN client, guardian will redirect correctly to the link that this was pointing to before this patch.