Load all l10n files
Previously, we'd only load `app.ftl`. With this change, all files
in a localisation directory will be loaded. This enables a couple
of things:

- Module-specific localisation files (e.g. separate files for
  the phones and bundle modules, which at the moment only need
  localisation to French and Spanish).
- Shared definitions, specifically, a brands.ftl file that can be
  referred to from the other Fluent files.

The main restriction at the moment is that I only implemented this
change for the frontend - Django still only loads app.ftl. I
figured this was fine, since the only strings added by Django are
the ones in wrapped_email.html, which don't change often and are
likely to remain in app.ftl. We can always update that to e.g. an
emails.ftl file, but we'd want to figure out how to avoid losing
attribution of specific strings to localisers.