Prevent line breaks in stats in Chinese
Although we want a stat label like "Trackers Removed" to span two
lines, we don't want every character in Chinese to be moved to a
new line. Although the former was achieved by setting a flex-basis
of 0, that caused line breaks between every character in Chinese.
By simply setting the flex-basis to a small (i.e. smaller than the
width of the word "Trackers") but non-zero (i.e. larger than the
width of 追) value, we achieve both goals.