Remove superfluous whitespace on mobile
I set a small flex-basis on stats to prevent line breaks in Chinese
for the stat, which was fine on desktop, but on mobile (where stats
are stacked vertically), it resulted in a bunch of whitespace at
the bottom.

This also prevents Chinese labels for slider positions (e.g.
"Promotionals") from having line breaks.