Remove unneeded arrow on block level description
When the description of the current block level is displayed
underneath the slider (i.e. on small screens), an arrow connects it
to the slider. On larger screens, however, it is to the side
and is connected through positioning, so the arrow should not be
visible (and, in fact, displayed as a floating rhombus).