Only check for a Relay number with phone masking
To avoid having to make a request to the /relaynumber/ API on every
page that might show the banner to ask people to fill in the phone
number launch survey, with this change we only call that when the
user actually has purchased phone masking.

The downside is that this means that the CSAT banner won't be shown
for users who have purchased phone masking but have not yet set up
their Relay number. This shouldn't happen often, when it happens
should only be a short period of time, and isn't that big of a
problem anyway, so that's an acceptable trade-off for improved
performance.