Add tips to the phone masking dashboard
Note that I had to move a bunch of things around from
phone.page.tsx to <PhoneDashboard>; although the tips should only
show on the dashboard (i.e. not on the phone onboarding), they
are shown on the far edge of the screen, whereas the rest of the
dashboard was wrapped in a container with a max-width. Thus, I
moved that container, as well as the "Resend welcome SMS" banner
(which is also only shown on the dashboard) into <PhoneDashboard>,
and then added the <Tips> component below that.