Move manage_flags blocking to the Serializer
Accessing serializer.data was invalid in the happy path; it's not
allowed to read that before validation. The `name` field had been
removed in serializer.validated_data, but you _can_ update the DB
row that would be updated using serializer.instance. Thus, on
updates, we can compare with serializer.instance.name.

However, instead of checking whether the user is attempting to
update `manage_flags` in the API endpoints, this commit also moves
it to where we parse user input as a flag in the first place.

Co-authored-by: John Whitlock <jwhitlock@mozilla.com>