Refactor return values when SMS text has no prefix
Instead of MatchByPrefix handling the case of no prefix, focus it on
when a prefix is found. This reduces the possbilities in MatchByPrefix
(match_type is full or short, prefix is non-nullable), reducing the
number of asserts to make mypy happy.

Rename _match_senders to _match_senders_by_prefix, to match the new
functionality.