Prevent subdomain explainer overflowing on mobile
On a small screen, it can get tight to have the explainer be a
popover attached to the indicator. Especially when the indicator is
in the middle of the viewport, there is not enough room for the
popover and all its content, so the content would break out of it.

This changes the popover into a dialog on small screens, so it can
overlap the indicator.