Consistent strings between pre- and first render
To avoid React complaining about mismatches between the HTML
rendered during the build, and the DOM it generates on the first
client-side render, we have to make sure we're using the same l10n
strings. Since we don't know the user's language at build-time,
this makes sure that we use the `en` strings on the first render
too. Immediately after the first render, we detect the user's
language and refresh the strings with the ones in their preferred
locales.