Allow trailing slashs on non-api_router API routes
Like API routes added by api_router, it is still allowed to omit
the trailing slash.

This works because by default, Django's APPEND_SLASH option is True
(see [1]), so if a route without a trailing slash it found, it will
append it, but not vice versa (i.e. a trailing slash will not be
automatically stripped).

[1] https://docs.djangoproject.com/en/dev/ref/middleware/#django.middleware.common.CommonMiddleware