Change TWILIO_MESSAGING_SERVICE_SID to CSV
Change the Twilio messaging service IDs settings to a comma-separated
list. Legacy values will be treated as a list on one entry.

When registeting a US relay phone mask with a messaging service, try
each service in order, stopping after the first successful service. Take
note of full services in the cache to avoid checking them.