Increase iat claim skew in test
With an iat claim skew of 6 seconds in the future, the test fails about
10% - 15% of the time, because test execution and rounding result in
the iat claim being only 5 seconds in the future.

Increase the iat claim skew, so that up to 5 seconds of test delay will
continue to raise the ImmatureSignatureError (although the test will
continue to fail if the test delay is more than 2 seconds).