Force the use of React 18 by all dependencies
Without that, our dependencies will be talking to different
versions of React, resulting in this error:
https://reactjs.org/warnings/invalid-hook-call-warning.html#duplicate-react

Unfortunately, I did have to blow away the entire previous lock
file, so it's imperative to keep an eye on things breaking. I
believe the newer version of npm changed its dependency resolution
algorithm, leading it to different versions installed with the
previous lock file. It's not great, and there might be a better
solution that I'm not aware of, but given that we're pretty good at
keeping our dependencies up-to-date, I think this is probably
relatively safe.