Mark two flaky tests as slow
The "can use masks and receive emails" test has been failing on
Firefox ever since our upgrade to Playwright 1.30.0 from 1.24.0:
https://github.com/mozilla/fx-private-relay/pull/3098

It's not clear what exactly in the upgrade would cause it to fail
consistently, but it was flaky already before that, and the other
test was too on Chromium. See e.g.
https://github.com/mozilla/fx-private-relay/actions/runs/4132141922/jobs/7140490726

Hopefully, giving them some more time to resolve should make them
pass more consistently. We could also specifically mark the browser
they are slow in in the callback, if we want. For more info, see
https://playwright.dev/docs/api/class-test#test-slow-1