Rewrite _handle_bounce to log more data
The counter "email_bounce" now has tags 'bounce_type', 'bounce_subtype',
'user', and 'relay_action'. The type and sub-type are lowercased.
The counter will be incremented for each recipient identified by AWS, or
once if no recipients.

The new log "bounce_notification" has the same data, except the type and
subtype are not lowercase. It also includes the data from the bounce
recipient 'bounce_action', 'bounce_status', and 'bounce_diagnostic'. If
there is more undocumented data, it is in 'bounce_extra'. There is also
'domain' for the user's domain.

The log "bounced recipient domain: {domain}" is still emitted, but
marked as legacy. It can be removed in Q4 2023, when we have a full
backlog of "bounce_notification" logs.

Some unused metrics have been removed:
* The counter "email_bounce_relay_user_gone" is replaced with the
  counter "email_bounce" with tag "user=missing".
* The counter "email_bounce_Permanent_General" is replaced with the
  counter "email_bounce" with tag "bounce_type=permanent",
  "bounce_subtype=general". Other similar tags that embedded the type
  and subtype in the name are replaced in the same way.

The handler no longer returns a 404 on the first missing user, but
continues processing the recipient list. This is not expected to have an
impact, since we forward emails to a single recipient.