Remove multiple scrollbars on Chrome
This was originally added in
66265f8e9623fa188a3d53fcb18dbbedf8ebc032, then removed in
https://github.com/mozilla/fx-private-relay/pull/3222, and then
was supposed to be restored (but only on large screens) in
https://github.com/mozilla/fx-private-relay/pull/3268 - but that
didn't happen. Thus, I've added it back here.
I also added a comment with more links to what happened in the
first place, to hopefully aid investigation into a proper fix down
the road.