Fix MIMEText content encoding
MIMEText expects a str, not bytes, for the first parameter. Our code was
working because email.Message.set_payload() handles bytes by decoding as
ascii with the 'surrogateescape' strategy from PEP 383.

For text messages, let MIMEText determine if the text is ASCII, and
fallback to UTF-8 if not. A plain ASCII text message is easier to see in
the raw message than base64-encoded UTF-8.

For HTML messages, our wrapper uses translations with placeholders,
which adds Unicode characters U+2068 and U+2069. Detection will always
pick UTF-8, so skip the extra work of detection.